day 25: fix subtraction with underflow in debug
I'm actually amazed this worked; q1 ended up negative, and yet somehow
q2 was still correct.

+++ day_25/src/main.rs
@@ -36,8 +36,8 @@ fn get_private_key(pubkey: u64) -> u64 {
         .map(|j| (transform(7, j), j))
         .collect::<HashMap<_, _>>();
 
-    let (_q1, q2) = extended_euclidean(20201227, transform(7, m));
-    let inverse = q2.rem_euclid(20201227);
+    let (_q1, q2) = extended_euclidean(20201227, transform(7, m) as i64);
+    let inverse = q2.rem_euclid(20201227) as u64;
 
     let mut y = pubkey;
     for i in 0..m {
@@ -50,11 +50,11 @@ fn get_private_key(pubkey: u64) -> u64 {
     panic!("could not find a suitable public key");
 }
 
-fn extended_euclidean(a: u64, b: u64) -> (u64, u64) {
+fn extended_euclidean(a: i64, b: i64) -> (i64, i64) {
     extended_euclidean_inner(a, b, 1, 0, 0, 1)
 }
 
-fn extended_euclidean_inner(a: u64, b: u64, s0: u64, s1: u64, t0: u64, t1: u64) -> (u64, u64) {
+fn extended_euclidean_inner(a: i64, b: i64, s0: i64, s1: i64, t0: i64, t1: i64) -> (i64, i64) {
     if b == 0 {
         assert_eq!(a, 1); // this algorithm makes no sense if gcd(orig_a, orig_b) != 1
         return (s0, t0);
